Replace short with int

+++ config.h
@@ -2,14 +2,14 @@ const static char *background_color = "#3e3e3e";
 const static char *border_color = "#ececec";
 const static char *font_color = "#ececec";
 const static char *font_pattern = "Inconsolata:style=Medium:size=15";
-const static unsigned short padding = 20;
+const static unsigned int padding = 20;
 
-const static unsigned short width = 300;
-const static unsigned short border_size = 2;
-const static unsigned short pos_x = 40;
-const static unsigned short pos_y = 50;
+const static unsigned int width = 300;
+const static unsigned int border_size = 2;
+const static unsigned int pos_x = 40;
+const static unsigned int pos_y = 50;
 
 enum corners { TOP_LEFT, TOP_RIGHT, BOTTOM_LEFT, BOTTOM_RIGHT };
 enum corners corner = TOP_RIGHT;
 
-const static unsigned short duration = 5; /* in seconds */
+const static unsigned int duration = 5; /* in seconds */
